Add files via upload
Added the functionality of a plinko traveling back into the board once it goes out of bounds. I've also changed how the velocity and plinko variables are stored, where both the x and y coordinates are stored in 2 variable array, which makes them easier to track. I realized later that storing the values in a class would be a better option. We were never taught how to use classes in python so I decided it wasn't worth redoing the code for it.  
Also added a second main plinkoTest function, which looks at the trajectory of the plinko from its initial position and alters the position and time step to get a more accurate result. It seems more accurate but not by much, it does collide with the pegs more however. 
We now need to add randomness to the code, energy loss and displaying velocity over time. hopefully it will make the collisions more accurate as it will be traveling slower. Looked at the lecture couldn't see how any of it related to collision detection, but it has a lot to do with the changing energy levels and frequencies.
If you have any questions just ask, it's late and I spent a lot more time on this than I thought, I thought I would be able to send this out sooner but I wanted to get more stuff working first. Collision should be good enough at this point. It's also less commented than my first piece of code but I hope its simple enough to follow